added login controller
